feat: implement bank account service

+++ src/main/java/com/sm/testsilkpay/model/web/TransferRequest.java
@@ -2,6 +2,8 @@ package com.sm.testsilkpay.model.web;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -9,8 +11,8 @@ import lombok.*;
 @AllArgsConstructor(staticName = "of")
 public class TransferRequest {
 
-  private Long from;
-  private Long to;
-  private Long amount;
+  private Long       from;
+  private Long       to;
+  private BigDecimal amount;
 
 }
